+++ packages/epub-codec/CHANGELOG.md
@@ -1,3 +1,10 @@
+## [1.3.3](https://github.com/ExaDev/documents.js/compare/epub-codec%401.3.2...epub-codec%401.3.3) (2026-09-08)
+
+
+### Dependencies
+
+- Updated document-schema.js to 7.5.1
+
 ## [1.3.2](https://github.com/ExaDev/documents.js/compare/epub-codec%401.3.1...epub-codec%401.3.2) (2026-09-08)